split mui route into per-component files
also lots of styling improvements
