--- apps/test-app/app/mui.tsx
@@ -1,549 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-import * as React from "react";
-import { type MetaFunction, useLocation } from "react-router";
-import Container from "@mui/material/Container";
-import IconButton from "@mui/material/IconButton";
-import Stack from "@mui/material/Stack";
-import Typography from "@mui/material/Typography";
-import visuallyHidden from "@mui/utils/visuallyHidden";
-import { Icon } from "@stratakit/mui";
-import * as NavigationList from "@stratakit/structures/unstable_NavigationList";
-import AccordionActions from "examples/mui/Accordion.actions.tsx";
-import AccordionDecoration from "examples/mui/Accordion.decoration.tsx";
-import AccordionDefault from "examples/mui/Accordion.default.tsx";
-import AccordionExpanded from "examples/mui/Accordion.expanded.tsx";
-import AccordionMarkerLeft from "examples/mui/Accordion.marker-left.tsx";
-import AccordionMultiple from "examples/mui/Accordion.multiple.tsx";
-import AccordionVariants from "examples/mui/Accordion.variants.tsx";
-import AlertPermutations_ from "examples/mui/Alert._permutations.tsx";
-import AlertDefault from "examples/mui/Alert.default.tsx";
-import AlertTitle from "examples/mui/Alert.title.tsx";
-import AppBarDefault from "examples/mui/AppBar.default.tsx";
-import AutocompleteDefault from "examples/mui/Autocomplete.default.tsx";
-import AutocompleteMultiple from "examples/mui/Autocomplete.multiple.tsx";
-import AutocompleteSizes from "examples/mui/Autocomplete.sizes.tsx";
-import AvatarDecorative from "examples/mui/Avatar.decorative.tsx";
-import AvatarDefault from "examples/mui/Avatar.default.tsx";
-import AvatarIcon from "examples/mui/Avatar.icon.tsx";
-import AvatarInitials from "examples/mui/Avatar.initials.tsx";
-import AvatarGroupDefault from "examples/mui/AvatarGroup.default.tsx";
-import BackdropDefault from "examples/mui/Backdrop.default.tsx";
-import BadgeColors from "examples/mui/Badge.colors.tsx";
-import BadgeDefault from "examples/mui/Badge.default.tsx";
-import BadgeError from "examples/mui/Badge.error.tsx";
-import BottomNavigationDefault from "examples/mui/BottomNavigation.default.tsx";
-import BreadcrumbsDefault from "examples/mui/Breadcrumbs.default.tsx";
-import ButtonIcons_ from "examples/mui/Button._icons.tsx";
-import ButtonPermutations_ from "examples/mui/Button._permutations.js";
-import ButtonColors from "examples/mui/Button.colors.tsx";
-import ButtonDefault from "examples/mui/Button.default.tsx";
-import ButtonSizes from "examples/mui/Button.sizes.tsx";
-import ButtonVariants from "examples/mui/Button.variants.tsx";
-import ButtonGroupDefault from "examples/mui/ButtonGroup.default.tsx";
-import ButtonGroupSizes from "examples/mui/ButtonGroup.sizes.tsx";
-import CardActions from "examples/mui/Card.actions.tsx";
-import CardDefault from "examples/mui/Card.default.tsx";
-import CardHeader from "examples/mui/Card.menu.tsx";
-import CheckboxChecked from "examples/mui/Checkbox.checked.tsx";
-import CheckboxDefault from "examples/mui/Checkbox.default.tsx";
-import CheckboxError from "examples/mui/Checkbox.error.tsx";
-import CheckboxGroup from "examples/mui/Checkbox.group.tsx";
-import CheckboxIndeterminate from "examples/mui/Checkbox.indeterminate.tsx";
-import ChipInteractive_ from "examples/mui/Chip._interactive.tsx";
-import ChipClickable from "examples/mui/Chip.clickable.tsx";
-import ChipDefault from "examples/mui/Chip.default.tsx";
-import ChipDeletable from "examples/mui/Chip.deletable.tsx";
-import ChipOutlined from "examples/mui/Chip.outlined.tsx";
-import ChipSizes from "examples/mui/Chip.sizes.tsx";
-import CircularProgressColors_ from "examples/mui/CircularProgress._colors.tsx";
-import CircularProgressDefault from "examples/mui/CircularProgress.default.tsx";
-import DialogDefault from "examples/mui/Dialog.default.tsx";
-import DividerDefault from "examples/mui/Divider.default.tsx";
-import DividerPresentational from "examples/mui/Divider.presentational.tsx";
-import DividerVertical from "examples/mui/Divider.vertical.tsx";
-import DrawerDefault from "examples/mui/Drawer.default.tsx";
-import FabDefault from "examples/mui/Fab.default.tsx";
-import IconButtonColors_ from "examples/mui/IconButton._colors.tsx";
-import IconButtonPlacements_ from "examples/mui/IconButton._placements.tsx";
-import IconButtonDefault from "examples/mui/IconButton.default.tsx";
-import IconButtonSizes from "examples/mui/IconButton.sizes.tsx";
-import LinearProgressColors_ from "examples/mui/LinearProgress._colors.tsx";
-import LinearProgressDefault from "examples/mui/LinearProgress.default.tsx";
-import LinkColors_ from "examples/mui/Link._colors.tsx";
-import LinkDefault from "examples/mui/Link.default.tsx";
-import ListAvatar from "examples/mui/List.avatar.tsx";
-import ListDefault from "examples/mui/List.default.tsx";
-import ListSubheader from "examples/mui/List.subheader.tsx";
-import MenuDefault from "examples/mui/Menu.default.tsx";
-import MenuDense from "examples/mui/Menu.dense.tsx";
-import MenuListDefault_ from "examples/mui/MenuList._default.tsx";
-import MobileStepperDefault from "examples/mui/MobileStepper.default.tsx";
-import NativeSelectDefault from "examples/mui/NativeSelect.default.tsx";
-import PaginationDefault from "examples/mui/Pagination.default.tsx";
-import PaperDefault from "examples/mui/Paper.default.tsx";
-import PopoverDefault from "examples/mui/Popover.default.tsx";
-import RadioGroupDefault from "examples/mui/RadioGroup.default.tsx";
-import RadioGroupError from "examples/mui/RadioGroup.error.tsx";
-import RatingDefault from "examples/mui/Rating.default.tsx";
-import SelectDefault from "examples/mui/Select.default.tsx";
-import SelectIcon from "examples/mui/Select.icon.tsx";
-import SelectMultiple from "examples/mui/Select.multiple.tsx";
-import SelectSizes from "examples/mui/Select.sizes.tsx";
-import SkeletonDefault from "examples/mui/Skeleton.default.tsx";
-import SkeletonVariants from "examples/mui/Skeleton.variants.tsx";
-import SliderDefault from "examples/mui/Slider.default.tsx";
-import SliderMarks from "examples/mui/Slider.marks.tsx";
-import SliderRange from "examples/mui/Slider.range.tsx";
-import SliderSizes from "examples/mui/Slider.sizes.tsx";
-import SliderTooltip from "examples/mui/Slider.tooltip.tsx";
-import SliderVertical from "examples/mui/Slider.vertical.tsx";
-import SnackbarDefault from "examples/mui/Snackbar.default.tsx";
-import SpeedDialDefault from "examples/mui/SpeedDial.default.tsx";
-import StepperClickable from "examples/mui/Stepper.clickable.tsx";
-import StepperDefault from "examples/mui/Stepper.default.tsx";
-import StepperOptional from "examples/mui/Stepper.optional.tsx";
-import SwipeableDrawerDefault from "examples/mui/SwipeableDrawer.default.tsx";
-import SwitchChecked from "examples/mui/Switch.checked.tsx";
-import SwitchDefault from "examples/mui/Switch.default.tsx";
-import SwitchSizes from "examples/mui/Switch.sizes.tsx";
-import TableDefault from "examples/mui/Table.default.tsx";
-import TableFooter from "examples/mui/Table.footer.tsx";
-import TableSelect from "examples/mui/Table.select.tsx";
-import TableSmall from "examples/mui/Table.small.tsx";
-import TabsDefault from "examples/mui/Tabs.default.tsx";
-import TabsScrollable from "examples/mui/Tabs.scrollable.tsx";
-import TextFieldDisabled from "examples/mui/TextField._disabled.tsx";
-import TextFieldDefault from "examples/mui/TextField.default.tsx";
-import TextFieldError from "examples/mui/TextField.error.tsx";
-import TextFieldIcon from "examples/mui/TextField.icon.tsx";
-import TextFieldMultiline from "examples/mui/TextField.multiline.tsx";
-import TextFieldSizes from "examples/mui/TextField.sizes.tsx";
-import ToggleButtonDisabled_ from "examples/mui/ToggleButton._disabled.tsx";
-import ToggleButtonPlacements_ from "examples/mui/ToggleButton._placements.tsx";
-import ToggleButtonDefault from "examples/mui/ToggleButton.default.tsx";
-import ToggleButtonSizes from "examples/mui/ToggleButton.sizes.tsx";
-import ToggleButtonStandalone from "examples/mui/ToggleButton.standalone.tsx";
-import ToggleButtonText from "examples/mui/ToggleButton.text.tsx";
-import TooltipDefault from "examples/mui/Tooltip.default.tsx";
-import TooltipDirection from "examples/mui/Tooltip.direction.tsx";
-import TypographyDefault from "examples/mui/Typography.default.tsx";
-import TypographyHeading from "examples/mui/Typography.heading.tsx";
-import TypographyVariants from "examples/mui/Typography.variants.tsx";
-import { SkipLinkContext } from "./~navigation.tsx";
-import { isProduction, useIsWideScreen } from "./~utils.tsx";
-
-import svgLink from "@stratakit/icons/link.svg";
-import styles from "./mui.module.css";
-
-// ----------------------------------------------------------------------------
-
-export const meta: MetaFunction = () => {
-	return [{ title: "StrataKit MUI theme" }];
-};
-
-// ----------------------------------------------------------------------------
-
-/**
- * Map of all MUI component examples displayed on this page.
- * Used to generate both the content and the navigation sidebar.
- */
-const components: Record<string, React.ReactNode> = {
-	Accordion: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<div>
-				<AccordionDefault />
-			</div>
-			<div>
-				<AccordionExpanded />
-			</div>
-			<div>
-				<AccordionMarkerLeft />
-			</div>
-			<div>
-				<AccordionDecoration />
-			</div>
-			<div>
-				<AccordionActions />
-			</div>
-			<AccordionMultiple />
-			<AccordionVariants />
-		</Stack>
-	),
-	Alert: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<AlertDefault />
-			<AlertTitle />
-			{!isProduction && <AlertPermutations_ />}
-		</Stack>
-	),
-	AppBar: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<AppBarDefault />
-		</Stack>
-	),
-	Autocomplete: (
-		<>
-			<AutocompleteDefault />
-			<AutocompleteSizes />
-			<AutocompleteMultiple />
-		</>
-	),
-	Avatar: (
-		<Stack spacing={1} direction="row">
-			<AvatarDefault />
-			<AvatarInitials />
-			<AvatarIcon />
-			<AvatarDecorative />
-		</Stack>
-	),
-	AvatarGroup: <AvatarGroupDefault />,
-	Backdrop: <BackdropDefault />,
-	Badge: (
-		<>
-			<BadgeDefault />
-			<BadgeColors />
-			<BadgeError />
-		</>
-	),
-	BottomNavigation: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<BottomNavigationDefault />
-		</Stack>
-	),
-	Breadcrumbs: <BreadcrumbsDefault />,
-	Button: (
-		<>
-			<ButtonDefault />
-			<ButtonSizes />
-			<ButtonVariants />
-			<ButtonColors />
-			{!isProduction && <ButtonIcons_ />}
-			{!isProduction && <ButtonPermutations_ />}
-		</>
-	),
-	ButtonGroup: (
-		<>
-			<ButtonGroupDefault />
-			<ButtonGroupSizes />
-		</>
-	),
-	Card: (
-		<>
-			<CardDefault />
-			<CardActions />
-			<CardHeader />
-		</>
-	),
-	Checkbox: (
-		<>
-			<CheckboxDefault />
-			<CheckboxChecked />
-			<CheckboxIndeterminate />
-			<CheckboxGroup />
-			<CheckboxError />
-		</>
-	),
-	Chip: (
-		<Stack spacing={1} direction="row" useFlexGap sx={{ flexWrap: "wrap" }}>
-			<ChipDefault />
-			<ChipOutlined />
-			<ChipClickable />
-			<ChipDeletable />
-			{!isProduction && <ChipInteractive_ />}
-			<ChipSizes />
-		</Stack>
-	),
-	CircularProgress: (
-		<>
-			<CircularProgressDefault />
-			{!isProduction && (
-				<Stack spacing={1} direction="row">
-					<CircularProgressColors_ />
-				</Stack>
-			)}
-		</>
-	),
-	Dialog: <DialogDefault />,
-	Divider: (
-		<>
-			<Stack sx={{ alignSelf: "stretch" }}>
-				<DividerDefault />
-			</Stack>
-			<DividerVertical />
-			<DividerPresentational />
-		</>
-	),
-	Drawer: <DrawerDefault />,
-	Fab: <FabDefault />,
-	IconButton: (
-		<>
-			<IconButtonDefault />
-			<IconButtonSizes />
-			{!isProduction && (
-				<Stack spacing={1} direction="row">
-					<IconButtonColors_ />
-				</Stack>
-			)}
-			{!isProduction && (
-				<Stack spacing={1} direction="row">
-					<IconButtonPlacements_ />
-				</Stack>
-			)}
-		</>
-	),
-	LinearProgress: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<LinearProgressDefault />
-			{!isProduction && <LinearProgressColors_ />}
-		</Stack>
-	),
-	Link: (
-		<>
-			<LinkDefault />
-			{!isProduction && (
-				<Stack spacing={1} direction="row">
-					<LinkColors_ />
-				</Stack>
-			)}
-		</>
-	),
-	List: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<ListDefault />
-			<ListAvatar />
-			<ListSubheader />
-		</Stack>
-	),
-	Menu: (
-		<>
-			<MenuDefault />
-			<MenuDense />
-		</>
-	),
-	...(isProduction ? {} : { MenuList: <MenuListDefault_ /> }),
-	MobileStepper: (
-		<Stack spacing={1} sx={{ alignSelf: "stretch" }}>
-			<MobileStepperDefault />
-		</Stack>
-	),
-	NativeSelect: <NativeSelectDefault />,
-	Pagination: <PaginationDefault />,
-	Paper: <PaperDefault />,
-	Popover: <PopoverDefault />,
-	RadioGroup: (
-		<>
-			<RadioGroupDefault />
-			<RadioGroupError />
-		</>
-	),
-	Rating: <RatingDefault />,
-	Select: (
-		<>
-			<SelectDefault />
-			<SelectIcon />
-			<SelectMultiple />
-			<SelectSizes />
-		</>
-	),
-	Skeleton: (
-		<Stack spacing={2} sx={{ alignSelf: "stretch" }}>
-			<SkeletonDefault />
-			<SkeletonVariants />
-		</Stack>
-	),
-	Slider: (
-		<>
-			<SliderDefault />
-			<SliderSizes />
-			<SliderTooltip />
-			<SliderMarks />
-			<SliderRange />
-			<SliderVertical />
-		</>
-	),
-	Snackbar: <SnackbarDefault />,
-	SpeedDial: <SpeedDialDefault />,
-	Stepper: (
-		<Stack spacing={4} sx={{ alignSelf: "stretch" }}>
-			<StepperDefault />
-			<StepperOptional />
-			<StepperClickable />
-		</Stack>
-	),
-	SwipeableDrawer: <SwipeableDrawerDefault />,
-	Switch: (
-		<>
-			<SwitchDefault />
-			<SwitchChecked />
-			<SwitchSizes />
-		</>
-	),
-	Table: (
-		<>
-			<TableDefault />
-			<TableSmall />
-			<TableSelect />
-			<TableFooter />
-		</>
-	),
-	Tabs: (
-		<>
-			<TabsDefault />
-			<TabsScrollable />
-		</>
-	),
-	TextField: (
-		<Stack spacing={1}>
-			<TextFieldDefault />
-			<TextFieldMultiline />
-			<TextFieldIcon />
-			<TextFieldError />
-			<TextFieldDisabled />
-			<TextFieldSizes />
-		</Stack>
-	),
-	ToggleButton: (
-		<>
-			<ToggleButtonDefault />
-			<ToggleButtonStandalone />
-			<ToggleButtonSizes />
-			<ToggleButtonDisabled_ />
-			<ToggleButtonText />
-			{!isProduction && (
-				<Stack spacing={1} direction="row">
-					<ToggleButtonPlacements_ />
-				</Stack>
-			)}
-		</>
-	),
-	Tooltip: (
-		<>
-			<TooltipDefault />
-			<TooltipDirection />
-		</>
-	),
-	Typography: (
-		<>
-			<TypographyDefault />
-			<TypographyHeading />
-			<TypographyVariants />
-		</>
-	),
-};
-
-// ----------------------------------------------------------------------------
-
-export default function Page() {
-	const location = useLocation();
-	const isWideScreen = useIsWideScreen();
-
-	// Scroll to heading on page load only.
-	React.useEffect(function scrollToHash() {
-		if (window.location.hash) {
-			const id = window.location.hash.slice(1);
-			const element = document.getElementById(id);
-			element?.scrollIntoView({ block: "start" });
-		}
-	}, []);
-
-	const navigationItems = Object.keys(components).map((name) => {
-		const id = name.toLowerCase().replace(" ", "-");
-		return (
-			<NavigationList.Anchor
-				key={id}
-				label={name}
-				href={`#${id}`}
-				active={location.hash === `#${id}`}
-			/>
-		);
-	});
-
-	return (
-		<div className={styles.page}>
-			{isWideScreen && (
-				<aside className={styles.sidebar}>
-					<Typography
-						variant="body1"
-						render={<h2 />}
-						sx={{ ml: 1, mb: 2, fontWeight: 500 }}
-					>
-						MUI components
-					</Typography>
-					<NavigationList.Root items={navigationItems} />
-				</aside>
-			)}
-
-			<Container
-				maxWidth="lg"
-				render={<main />}
-				className={styles.main}
-				tabIndex={-1}
-				id={React.use(SkipLinkContext)?.id}
-			>
-				<Typography variant="h4" render={<h1 />} className={styles.h1}>
-					StrataKit MUI theme
-				</Typography>
-
-				{Object.entries(components).map(([name, content]) => (
-					<ComponentExamples key={name} name={name}>
-						{content}
-					</ComponentExamples>
-				))}
-			</Container>
-		</div>
-	);
-}
-
-// ----------------------------------------------------------------------------
-
-interface ComponentExamplesProps {
-	name: string;
-	children?: React.ReactNode;
-}
-
-function ComponentExamples(props: ComponentExamplesProps) {
-	const { name } = props;
-	const id = name.toLowerCase().replace(" ", "-");
-	const isTarget = useLocation().hash === `#${id}`;
-
-	return (
-		<div
-			className={styles.exampleContainer}
-			data-target={isTarget ? "true" : undefined}
-		>
-			<hgroup className={styles.exampleHeader}>
-				<Typography
-					variant="h5"
-					render={<h2 />}
-					id={id}
-					className={styles.exampleTitle}
-					tabIndex={-1}
-				>
-					{name}
-				</Typography>
-				<IconButton
-					render={<a />}
-					id={`${id}-permalink`}
-					aria-labelledby={`${id}-permalink ${id}`}
-					className={styles.examplePermalink}
-					href={`#${id}`}
-				>
-					<Icon href={svgLink} />
-					<span style={visuallyHidden}>Permalink</span>
-				</IconButton>
-			</hgroup>
-
-			<Stack
-				spacing={2}
-				sx={{ alignItems: "start" }}
-				className={styles.exampleContent}
-			>
-				{props.children}
-			</Stack>
-		</div>
-	);
-}
