+++ apps/test-app/app/mui/AvatarGroup.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import AvatarGroupDefault from "examples/mui/AvatarGroup.default.tsx";
+
+export default function AvatarGroupExamples() {
+	return <AvatarGroupDefault />;
+}
