+++ apps/test-app/app/mui/Backdrop.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import BackdropDefault from "examples/mui/Backdrop.default.tsx";
+
+export default function BackdropExamples() {
+	return <BackdropDefault />;
+}
