+++ apps/test-app/app/mui/Breadcrumbs.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import BreadcrumbsDefault from "examples/mui/Breadcrumbs.default.tsx";
+
+export default function BreadcrumbsExamples() {
+	return <BreadcrumbsDefault />;
+}
