+++ apps/test-app/app/mui/ButtonGroup.tsx
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import ButtonGroupDefault from "examples/mui/ButtonGroup.default.tsx";
+import ButtonGroupSizes from "examples/mui/ButtonGroup.sizes.tsx";
+
+export default function ButtonGroupExamples() {
+	return (
+		<>
+			<ButtonGroupDefault />
+			<ButtonGroupSizes />
+		</>
+	);
+}
