+++ apps/test-app/app/mui/Card.tsx
@@ -1,0 +1,17 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import CardActions from "examples/mui/Card.actions.tsx";
+import CardDefault from "examples/mui/Card.default.tsx";
+import CardHeader from "examples/mui/Card.menu.tsx";
+
+export default function CardExamples() {
+	return (
+		<>
+			<CardDefault />
+			<CardActions />
+			<CardHeader />
+		</>
+	);
+}
