+++ apps/test-app/app/mui/Dialog.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import DialogDefault from "examples/mui/Dialog.default.tsx";
+
+export default function DialogExamples() {
+	return <DialogDefault />;
+}
