+++ apps/test-app/app/mui/Drawer.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import DrawerDefault from "examples/mui/Drawer.default.tsx";
+
+export default function DrawerExamples() {
+	return <DrawerDefault />;
+}
