+++ apps/test-app/app/mui/Fab.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import FabDefault from "examples/mui/Fab.default.tsx";
+
+export default function FabExamples() {
+	return <FabDefault />;
+}
