+++ apps/test-app/app/mui/NativeSelect.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import NativeSelectDefault from "examples/mui/NativeSelect.default.tsx";
+
+export default function NativeSelectExamples() {
+	return <NativeSelectDefault />;
+}
