+++ apps/test-app/app/mui/Pagination.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import PaginationDefault from "examples/mui/Pagination.default.tsx";
+
+export default function PaginationExamples() {
+	return <PaginationDefault />;
+}
