+++ apps/test-app/app/mui/Paper.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import PaperDefault from "examples/mui/Paper.default.tsx";
+
+export default function PaperExamples() {
+	return <PaperDefault />;
+}
