+++ apps/test-app/app/mui/Popover.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import PopoverDefault from "examples/mui/Popover.default.tsx";
+
+export default function PopoverExamples() {
+	return <PopoverDefault />;
+}
