+++ apps/test-app/app/mui/Rating.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import RatingDefault from "examples/mui/Rating.default.tsx";
+
+export default function RatingExamples() {
+	return <RatingDefault />;
+}
