+++ apps/test-app/app/mui/Select.tsx
@@ -1,0 +1,19 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import SelectDefault from "examples/mui/Select.default.tsx";
+import SelectIcon from "examples/mui/Select.icon.tsx";
+import SelectMultiple from "examples/mui/Select.multiple.tsx";
+import SelectSizes from "examples/mui/Select.sizes.tsx";
+
+export default function SelectExamples() {
+	return (
+		<>
+			<SelectDefault />
+			<SelectIcon />
+			<SelectMultiple />
+			<SelectSizes />
+		</>
+	);
+}
