+++ apps/test-app/app/mui/Snackbar.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import SnackbarDefault from "examples/mui/Snackbar.default.tsx";
+
+export default function SnackbarExamples() {
+	return <SnackbarDefault />;
+}
