+++ apps/test-app/app/mui/SpeedDial.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import SpeedDialDefault from "examples/mui/SpeedDial.default.tsx";
+
+export default function SpeedDialExamples() {
+	return <SpeedDialDefault />;
+}
