+++ apps/test-app/app/mui/SwipeableDrawer.tsx
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import SwipeableDrawerDefault from "examples/mui/SwipeableDrawer.default.tsx";
+
+export default function SwipeableDrawerExamples() {
+	return <SwipeableDrawerDefault />;
+}
