+++ apps/test-app/app/mui/Switch.tsx
@@ -1,0 +1,17 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import SwitchChecked from "examples/mui/Switch.checked.tsx";
+import SwitchDefault from "examples/mui/Switch.default.tsx";
+import SwitchSizes from "examples/mui/Switch.sizes.tsx";
+
+export default function SwitchExamples() {
+	return (
+		<>
+			<SwitchDefault />
+			<SwitchChecked />
+			<SwitchSizes />
+		</>
+	);
+}
