+++ apps/test-app/app/mui/Tabs.tsx
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import TabsDefault from "examples/mui/Tabs.default.tsx";
+import TabsScrollable from "examples/mui/Tabs.scrollable.tsx";
+
+export default function TabsExamples() {
+	return (
+		<>
+			<TabsDefault />
+			<TabsScrollable />
+		</>
+	);
+}
