+++ apps/test-app/app/mui/ToggleButton.tsx
@@ -1,0 +1,29 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import Stack from "@mui/material/Stack";
+import ToggleButtonDisabled_ from "examples/mui/ToggleButton._disabled.tsx";
+import ToggleButtonPlacements_ from "examples/mui/ToggleButton._placements.tsx";
+import ToggleButtonDefault from "examples/mui/ToggleButton.default.tsx";
+import ToggleButtonSizes from "examples/mui/ToggleButton.sizes.tsx";
+import ToggleButtonStandalone from "examples/mui/ToggleButton.standalone.tsx";
+import ToggleButtonText from "examples/mui/ToggleButton.text.tsx";
+import { isProduction } from "../~utils.tsx";
+
+export default function ToggleButtonExamples() {
+	return (
+		<>
+			<ToggleButtonDefault />
+			<ToggleButtonStandalone />
+			<ToggleButtonSizes />
+			<ToggleButtonDisabled_ />
+			<ToggleButtonText />
+			{!isProduction && (
+				<Stack spacing={1} direction="row">
+					<ToggleButtonPlacements_ />
+				</Stack>
+			)}
+		</>
+	);
+}
