+++ apps/test-app/app/mui/Tooltip.tsx
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import TooltipDefault from "examples/mui/Tooltip.default.tsx";
+import TooltipDirection from "examples/mui/Tooltip.direction.tsx";
+
+export default function TooltipExamples() {
+	return (
+		<>
+			<TooltipDefault />
+			<TooltipDirection />
+		</>
+	);
+}
