+++ apps/test-app/app/mui/mui.module.css
@@ -8,7 +8,10 @@
 	display: flex;
 	block-size: 100%;
 
+	background-color: var(--stratakit-color-bg-page-depth);
+
 	:where(html:has(&)) {
+		scrollbar-gutter: stable;
 		scroll-padding-block: var(--stratakit-space-x7);
 
 		@media (prefers-reduced-motion: no-preference) {
@@ -54,7 +57,7 @@
 }
 
 .examplePermalink {
-	:where(.exampleHeader:is(:not(:hover, :focus-within))) & {
+	:where(.exampleContainer:is(:not(:hover, :focus-within, [data-target]))) & {
 		opacity: 0;
 	}
 }
@@ -62,7 +65,7 @@
 .exampleContainer {
 	display: flex;
 	flex-direction: column;
-	gap: var(--stratakit-space-x4);
+	gap: var(--stratakit-space-x2);
 	position: relative;
 
 	&:where(:not(:nth-last-child(1 of &))) {
@@ -71,14 +74,20 @@
 
 	&:where([data-target]) {
 		scroll-margin-block: var(--stratakit-space-x7);
+	}
+}
 
-		&::before {
-			content: "";
-			position: absolute;
-			inset: calc(-1 * var(--stratakit-space-x4));
-			outline: 1px dashed var(--stratakit-color-border-neutral-base);
-			border-radius: 4px;
-			pointer-events: none;
-		}
+.exampleContent {
+	position: relative;
+
+	background-color: var(--stratakit-color-bg-page-base);
+	border: 1px solid var(--stratakit-color-border-neutral-muted);
+
+	padding: var(--stratakit-space-x7);
+	border-radius: 8px;
+
+	:where(.exampleContainer[data-target]) & {
+		outline: 1px solid var(--stratakit-color-border-mono-base);
+		outline-offset: 1px;
 	}
 }
