+++ apps/test-app/app/routes.ts
@@ -13,7 +13,7 @@ export default [
 	route("sandbox", "./sandbox/index.tsx"),
 	route("tokens", "./tokens.tsx"),
 	route("icons", "./icons.tsx"),
-	route("mui", "./mui.tsx"),
+	route("mui", "./mui/mui.tsx"),
 
 	layout(
 		"./tests/tests.tsx",
